[UI, Refactor] Removed the View menu (also removed code related to it). [Bugfix] Nodes can now be properly moved when the network's Zoom is different than multiples of 100% (This could be done in the past, but the movement of the nodes didn't represent that of the user). [Refactor, Tests] IconLoader has been replaced for IconBind to offer stability regarding available icons. A test now verifies every Icon is bound to an image file that can be loaded. [Refactor] Changed shortcuts and added new ones (Ctrl+Alt+S: Save and reopen. Ctrl+Shift+S: Save as. Ctrl+D: Network properties. Ctrl+W: Close current tab. Ctrl+Q: Quit OpenMarkov. Ctrl+Shift+Left Arrow: Go to previous tab. Ctrl+Shift+Right Arrow: Go to next tab). [UX] The 'Change working mode' button is now only enabled for Influence Diagrams and Bayesian Networks. [UX] The zoom buttons and combobox are now only enabled when a tab is opened. [Refactor] Removed 'last connection' and 'last user connected' preferences as they were never used. [UX, Refactor] Replaced 'Switch to inference mode' and 'Switch to edit mode' for a single button. Depending on the current working state, this button shows 'Switch to inference mode' or 'Switch to edit mode'. [Feature] Preferences now can be stored in a directory inside the User's home directory. [Refactor] Changed the preference preferred order to 'Home dir - Installation dir - Backing Store - Session'. [Refactor, Feature] Preferences are now stored in JSON format. [Refactor, UX] Enclosing quotes can now be used on filepaths in Windows System.

+++ src/main/java/org/openmarkov/gui/graphic/VisualNode.java
@@ -106,12 +106,12 @@ public abstract class VisualNode extends VisualElement implements ClassLocalizab
     /**
      * Value of the X coordinate in temporal position of the node.
      */
-    protected int temporalCoordinateX;
+    protected double temporalCoordinateX;
     
     /**
      * Value of the Y coordinate in temporal position of the node.
      */
-    protected int temporalCoordinateY;
+    protected double temporalCoordinateY;
     
     public VisualNode(Node node, VisualNetwork visualNetwork) {
         this.node = node;
@@ -182,9 +182,8 @@ public abstract class VisualNode extends VisualElement implements ClassLocalizab
      * @param value new position.
      */
     public void setTemporalPosition(Point2D.Double value) {
-        
-        temporalCoordinateX = (int) value.getX();
-        temporalCoordinateY = (int) value.getY();
+        temporalCoordinateX = value.getX();
+        temporalCoordinateY = value.getY();
     }
     
     /**
